+++ addon/components/snippet-plugin/snippet-list/snippet-list-view.ts
@@ -1,10 +1,13 @@
 import Component from '@glimmer/component';
 import { action } from '@ember/object';
+import { SnippetList } from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/snippet-plugin';
 
 interface Args {
+  snippetLists: SnippetList[];
   assignedSnippetListsIds: string[];
+  listNameFilter: string | null;
+  isLoading: boolean;
   onChange: (assignedSnippetListsIds: string[]) => void;
-  closeModal: () => void;
 }
 
 export default class SnippetListViewComponent extends Component<Args> {
@@ -24,5 +27,35 @@ export default class SnippetListViewComponent extends Component<Args> {
     );
 
     return this.args.onChange(newSnippetListIds);
+  }
+
+  @action
+  onClickRow(snippetList: SnippetList, event: Event) {
+    if (!(event.target instanceof HTMLElement)) {
+      return;
+    }
+
+    const target = event.target;
+
+    // Only trigger the action when clicking on a td or tr element
+    // This is to prevent the action from being triggered when clicking on a checkbox
+    if (target.tagName !== 'TD' && target.tagName !== 'TR') {
+      return;
+    }
+
+    const snippetListId = snippetList.id;
+
+    if (!snippetListId) {
+      return;
+    }
+
+    const isSelected =
+      this.args.assignedSnippetListsIds.includes(snippetListId);
+
+    this.onChange(snippetListId, !isSelected);
+  }
+
+  get snippetLists() {
+    return this.args.snippetLists;
   }
 }